Separate folder for sorting algorithms
